Move out nested classes from SdkHarnessClient.
The SdkHarnessClient contained some nested subclasses complex
enough to merit their own files.  This commit hopefully simplifies
its reading by separating out related but distinct concepts into
their own class files.
