+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/control/SdkHarnessClient.java
@@ -20,28 +20,19 @@ package org.apache.beam.runners.fnexecution.control;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
 
-import com.google.auto.value.AutoValue;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.InstructionResponse;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.ProcessBundleDescriptor;
-import org.apache.beam.model.fnexecution.v1.BeamFnApi.ProcessBundleRequest;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.RegisterResponse;
 import org.apache.beam.model.pipeline.v1.Endpoints;
 import org.apache.beam.runners.fnexecution.data.FnDataService;
 import org.apache.beam.runners.fnexecution.data.RemoteInputDestination;
 import org.apache.beam.runners.fnexecution.state.StateDelegator;
 import org.apache.beam.runners.fnexecution.state.StateRequestHandler;
-import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.fn.data.CloseableFnDataReceiver;
-import org.apache.beam.sdk.fn.data.FnDataReceiver;
-import org.apache.beam.sdk.fn.data.InboundDataClient;
-import org.apache.beam.sdk.fn.data.LogicalEndpoint;
-import org.apache.beam.sdk.util.MoreFutures;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -70,229 +61,6 @@ public class SdkHarnessClient implements AutoCloseable {
     @Override
     public String getId() {
       return String.valueOf(nextId.incrementAndGet());
-    }
-  }
-
-  /**
-   * A processor capable of creating bundles for some registered {@link ProcessBundleDescriptor}.
-   */
-  public class BundleProcessor<T> {
-    private final ProcessBundleDescriptor processBundleDescriptor;
-    private final CompletionStage<RegisterResponse> registrationFuture;
-    private final RemoteInputDestination<WindowedValue<T>> remoteInput;
-    private final StateDelegator stateDelegator;
-
-    private BundleProcessor(
-        ProcessBundleDescriptor processBundleDescriptor,
-        CompletionStage<RegisterResponse> registrationFuture,
-        RemoteInputDestination<WindowedValue<T>> remoteInput,
-        StateDelegator stateDelegator) {
-      this.processBundleDescriptor = processBundleDescriptor;
-      this.registrationFuture = registrationFuture;
-      this.remoteInput = remoteInput;
-      this.stateDelegator = stateDelegator;
-    }
-
-    public CompletionStage<RegisterResponse> getRegistrationFuture() {
-      return registrationFuture;
-    }
-
-    /**
-     * Start a new bundle for the given {@link BeamFnApi.ProcessBundleDescriptor} identifier.
-     *
-     * <p>The input channels for the returned {@link ActiveBundle} are derived from the instructions
-     * in the {@link BeamFnApi.ProcessBundleDescriptor}.
-     *
-     * <p>NOTE: It is important to {@link #close()} each bundle after all elements are emitted.
-     * <pre>{@code
-     * try (ActiveBundle<InputT> bundle = SdkHarnessClient.newBundle(...)) {
-     *   FnDataReceiver<InputT> inputReceiver = bundle.getInputReceiver();
-     *   // send all elements ...
-     * }
-     * }</pre>
-     */
-    public ActiveBundle<T> newBundle(
-        Map<BeamFnApi.Target, RemoteOutputReceiver<?>> outputReceivers) {
-      return newBundle(outputReceivers, request -> {
-        throw new UnsupportedOperationException(String.format(
-            "The %s does not have a registered state handler.",
-            ActiveBundle.class.getSimpleName()));
-      });
-    }
-
-    /**
-     * Start a new bundle for the given {@link BeamFnApi.ProcessBundleDescriptor} identifier.
-     *
-     * <p>The input channels for the returned {@link ActiveBundle} are derived from the instructions
-     * in the {@link BeamFnApi.ProcessBundleDescriptor}.
-     *
-     * <p>NOTE: It is important to {@link #close()} each bundle after all elements are emitted.
-     * <pre>{@code
-     * try (ActiveBundle<InputT> bundle = SdkHarnessClient.newBundle(...)) {
-     *   FnDataReceiver<InputT> inputReceiver = bundle.getInputReceiver();
-     *   // send all elements ...
-     * }
-     * }</pre>
-     */
-    public ActiveBundle<T> newBundle(
-        Map<BeamFnApi.Target, RemoteOutputReceiver<?>> outputReceivers,
-        StateRequestHandler stateRequestHandler) {
-      String bundleId = idGenerator.getId();
-
-      final CompletionStage<BeamFnApi.InstructionResponse> genericResponse =
-          fnApiControlClient.handle(
-              BeamFnApi.InstructionRequest.newBuilder()
-                  .setInstructionId(bundleId)
-                  .setProcessBundle(
-                      BeamFnApi.ProcessBundleRequest.newBuilder()
-                          .setProcessBundleDescriptorReference(processBundleDescriptor.getId()))
-                  .build());
-      LOG.debug(
-          "Sent {} with ID {} for {} with ID {}",
-          ProcessBundleRequest.class.getSimpleName(),
-          bundleId,
-          ProcessBundleDescriptor.class.getSimpleName(),
-          processBundleDescriptor.getId());
-
-      CompletionStage<BeamFnApi.ProcessBundleResponse> specificResponse =
-          genericResponse.thenApply(InstructionResponse::getProcessBundle);
-      Map<BeamFnApi.Target, InboundDataClient> outputClients = new HashMap<>();
-      for (Map.Entry<BeamFnApi.Target, RemoteOutputReceiver<?>> targetReceiver
-          : outputReceivers.entrySet()) {
-        InboundDataClient outputClient =
-            attachReceiver(
-                bundleId,
-                targetReceiver.getKey(),
-                (RemoteOutputReceiver) targetReceiver.getValue());
-        outputClients.put(targetReceiver.getKey(), outputClient);
-      }
-
-      CloseableFnDataReceiver<WindowedValue<T>> dataReceiver =
-          fnApiDataService.send(
-              LogicalEndpoint.of(bundleId, remoteInput.getTarget()), remoteInput.getCoder());
-
-      return new ActiveBundle(
-          bundleId,
-          specificResponse,
-          dataReceiver,
-          outputClients,
-          stateDelegator.registerForProcessBundleInstructionId(bundleId, stateRequestHandler));
-    }
-
-    private <OutputT> InboundDataClient attachReceiver(
-        String bundleId,
-        BeamFnApi.Target target,
-        RemoteOutputReceiver<WindowedValue<OutputT>> receiver) {
-      return fnApiDataService.receive(
-          LogicalEndpoint.of(bundleId, target), receiver.getCoder(), receiver.getReceiver());
-    }
-  }
-
-  /** An active bundle for a particular {@link BeamFnApi.ProcessBundleDescriptor}. */
-  public static class ActiveBundle<InputT> implements AutoCloseable {
-    private final String bundleId;
-    private final CompletionStage<BeamFnApi.ProcessBundleResponse> response;
-    private final CloseableFnDataReceiver<WindowedValue<InputT>> inputReceiver;
-    private final Map<BeamFnApi.Target, InboundDataClient> outputClients;
-    private final StateDelegator.Registration stateRegistration;
-
-    private ActiveBundle(
-        String bundleId,
-        CompletionStage<BeamFnApi.ProcessBundleResponse> response,
-        CloseableFnDataReceiver<WindowedValue<InputT>> inputReceiver,
-        Map<BeamFnApi.Target, InboundDataClient> outputClients,
-        StateDelegator.Registration stateRegistration) {
-      this.bundleId = bundleId;
-      this.response = response;
-      this.inputReceiver = inputReceiver;
-      this.outputClients = outputClients;
-      this.stateRegistration = stateRegistration;
-    }
-
-    /**
-     * Returns an id used to represent this bundle.
-     */
-    public String getBundleId() {
-      return bundleId;
-    }
-
-    /**
-     * Returns a {@link FnDataReceiver receiver} which consumes input elements forwarding them
-     * to the SDK.
-     */
-    public FnDataReceiver<WindowedValue<InputT>> getInputReceiver() {
-      return inputReceiver;
-    }
-
-    /**
-     * Blocks till bundle processing is finished. This is comprised of:
-     * <ul>
-     *   <li>closing the {@link #getInputReceiver() input receiver}.</li>
-     *   <li>waiting for the SDK to say that processing the bundle is finished.</li>
-     *   <li>waiting for all inbound data clients to complete</li>
-     * </ul>
-     *
-     * <p>This method will throw an exception if bundle processing has failed.
-     * {@link Throwable#getSuppressed()} will return all the reasons as to why processing has
-     * failed.
-     */
-    @Override
-    public void close() throws Exception {
-      Exception exception = null;
-      try {
-        inputReceiver.close();
-      } catch (Exception e) {
-        exception = e;
-      }
-      try {
-        // We don't have to worry about the completion stage.
-        if (exception == null) {
-          MoreFutures.get(response);
-        } else {
-          // TODO: [BEAM-3962] Handle aborting the bundle being processed.
-          throw new IllegalStateException("Processing bundle failed, "
-              + "TODO: [BEAM-3962] abort bundle.");
-        }
-      } catch (Exception e) {
-        if (exception == null) {
-          exception = e;
-        } else {
-          exception.addSuppressed(e);
-        }
-      }
-      try {
-        if (exception == null) {
-          stateRegistration.deregister();
-        } else {
-          stateRegistration.abort();
-        }
-      } catch (Exception e) {
-        if (exception == null) {
-          exception = e;
-        } else {
-          exception.addSuppressed(e);
-        }
-      }
-      for (InboundDataClient outputClient : outputClients.values()) {
-        try {
-          // If we failed processing this bundle, we should cancel all inbound data.
-          if (exception == null) {
-            outputClient.awaitCompletion();
-          } else {
-            outputClient.cancel();
-          }
-        } catch (Exception e) {
-          if (exception == null) {
-            exception = e;
-          } else {
-            exception.addSuppressed(e);
-          }
-        }
-      }
-      if (exception != null) {
-        throw exception;
-      }
-
     }
   }
 
@@ -403,7 +171,7 @@ public class SdkHarnessClient implements AutoCloseable {
                 descriptor,
                 (RemoteInputDestination) remoteInputDesination,
                 stateDelegator));
-    checkArgument(bundleProcessor.processBundleDescriptor.equals(descriptor),
+    checkArgument(bundleProcessor.getProcessBundleDescriptor().equals(descriptor),
         "The provided %s with id %s collides with an existing %s with the same id but "
             + "containing different contents.",
         BeamFnApi.ProcessBundleDescriptor.class.getSimpleName(),
@@ -436,7 +204,7 @@ public class SdkHarnessClient implements AutoCloseable {
         genericResponse.thenApply(InstructionResponse::getRegister);
 
     BundleProcessor<T> bundleProcessor = new BundleProcessor<>(
-        processBundleDescriptor,
+        fnApiControlClient, fnApiDataService, idGenerator, processBundleDescriptor,
         registerResponseFuture,
         remoteInputDestination,
         stateDelegator);
@@ -447,17 +215,4 @@ public class SdkHarnessClient implements AutoCloseable {
   @Override
   public void close() {}
 
-  /**
-   * A pair of {@link Coder} and {@link FnDataReceiver} which can be registered to receive elements
-   * for a {@link LogicalEndpoint}.
-   */
-  @AutoValue
-  public abstract static class RemoteOutputReceiver<T> {
-    public static <T> RemoteOutputReceiver of (Coder<T> coder, FnDataReceiver<T> receiver) {
-      return new AutoValue_SdkHarnessClient_RemoteOutputReceiver<>(coder, receiver);
-    }
-
-    public abstract Coder<T> getCoder();
-    public abstract FnDataReceiver<T> getReceiver();
-  }
 }
